Bench: Use different endpoint for /config
This change moves the configuration endpoint to a sandbox/ endpoint
as well as moving the handler function to sandbox.go.
