+++ cmd/oceanbench/main.go
@@ -215,7 +215,6 @@ func main() {
 	http.HandleFunc("/set/devices/edit/sensor", editSensorHandler)
 	http.HandleFunc("/set/devices/edit/actuator", editActuatorHandler)
 	http.HandleFunc("/set/devices/edit", editDevicesHandler)
-	http.HandleFunc("/set/devices/configure", configDevicesHandler)
 	http.HandleFunc("/set/devices/", setDevicesHandler)
 	http.HandleFunc("/set/crons/edit", editCronsHandler)
 	http.HandleFunc("/set/crons/", setCronsHandler)
@@ -236,6 +235,7 @@ func main() {
 	http.HandleFunc("/admin/user/delete", adminHandler)
 	http.HandleFunc("/admin/site", adminHandler)
 	http.HandleFunc("/admin/broadcast", adminHandler)
+	http.HandleFunc("/admin/sandbox/configure", configDevicesHandler)
 	http.HandleFunc("/admin/utils", adminHandler)
 	http.HandleFunc("/data/", dataHandler)
 	http.HandleFunc("/throughputs", throughputsHandler)
